+++ solidity/test/bridge/Bridge.Deposit.test.ts
@@ -45,7 +45,7 @@ describe("Bridge - Deposit", () => {
   let relay: FakeContract<IRelay>
   let bridge: Bridge & BridgeStub
   let bridgeGovernance: BridgeGovernance
-  let deployBridge: (txProofDifficultyFactor: number) => Promise<Contract>
+  let deployBridge: (bridgeType: string, txProofDifficultyFactor: number) => Promise<Contract>
 
   before(async () => {
     // eslint-disable-next-line @typescript-eslint/no-extra-semi
@@ -2570,7 +2570,7 @@ describe("Bridge - Deposit", () => {
         context(
           "when accumulated difficulty in headers chain is insufficient",
           () => {
-            let otherBridge: Bridge & BridgeStub
+            let otherBridge: Bridge
             const data: DepositSweepTestData = JSON.parse(
               JSON.stringify(SingleP2SHDeposit)
             )
@@ -2598,7 +2598,7 @@ describe("Bridge - Deposit", () => {
               // to deem transaction proof validity. This scenario uses test
               // data which has only 6 confirmations. That should force the
               // failure we expect within this scenario.
-              otherBridge = (await deployBridge(12)) as BridgeStub
+              otherBridge = (await deployBridge("Bridge", 12)) as Bridge
               await otherBridge.setSpvMaintainerStatus(
                 spvMaintainer.address,
                 true
