+++ solidity/test/bridge/Bridge.Redemption.test.ts
@@ -58,7 +58,7 @@ describe("Bridge - Redemption", () => {
   let bridgeGovernance: BridgeGovernance
   let walletRegistry: FakeContract<IWalletRegistry>
 
-  let deployBridge: (txProofDifficultyFactor: number) => Promise<Contract>
+  let deployBridge: (bridgeType: string, txProofDifficultyFactor: number) => Promise<Contract>
 
   let redemptionTimeout: number
   let redemptionTimeoutSlashingAmount: BigNumber
@@ -3504,7 +3504,7 @@ describe("Bridge - Redemption", () => {
       context(
         "when accumulated difficulty in headers chain is insufficient",
         () => {
-          let otherBridge: Bridge & BridgeStub
+          let otherBridge: Bridge
           const data: RedemptionTestData = JSON.parse(
             JSON.stringify(SinglePendingRequestedRedemption)
           )
@@ -3521,7 +3521,7 @@ describe("Bridge - Redemption", () => {
             // to deem transaction proof validity. This scenario uses test
             // data which has only 6 confirmations. That should force the
             // failure we expect within this scenario.
-            otherBridge = (await deployBridge(12)) as BridgeStub
+            otherBridge = (await deployBridge("Bridge", 12)) as Bridge
             await otherBridge.setSpvMaintainerStatus(
               spvMaintainer.address,
               true
